separate creatorelectiondata in three class to correspond from each file
